test(scheduler): split spawn event coverage
